+++ components/AnnouncementBar/index.tsx
