+++ components/EmojiLayer/emoji-list-data.json
