+++ components/EmojiLayer/index.tsx
