+++ components/Footer/footer-data.ts
