+++ middleware/index.ts
@@ -1,3 +1,2 @@
 export * from "./hackathons-api";
 export * from "./teams-api";
-export * from "./utils";
