+++ pages/events/index.tsx
@@ -5,11 +5,11 @@ import Link from "next/link";
 import { NextPage } from "next";
 
 // Components Imports
-import Callout from "@/common/components/Callout";
-import OrganizeWithUs from "@/common/components/OrganizeWithUs";
-import Button from "@/common/components/ui-patterns/Button";
-import UpcomingHackathons from "@/common/components/UpcomingHackathons";
-import UpcomingSessions from "@/common/components/UpcomingSessions";
+import Callout from "@/components/Callout";
+import OrganizeWithUs from "@/components/OrganizeWithUs";
+import Button from "@/components/ui-patterns/Button";
+import UpcomingHackathons from "@/components/UpcomingHackathons";
+import UpcomingSessions from "@/components/UpcomingSessions";
 
 const Events: NextPage = () => {
   return (
