+++ pages/mlh.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { NextPage } from "next";
 
 // Components Imports
-import Callout from "@/common/components/Callout";
+import Callout from "@/components/Callout";
 
 const MLH: NextPage = () => {
   return (
